fix: remove events exports from server
