+++ packages/shared/src/types.ts
@@ -1,49 +1,5 @@
 export type PrimitiveValue = null | boolean | string | number;
 
-export enum ProviderEvents {
-  /**
-   * The provider is ready to evaluate flags.
-   */
-  Ready = 'PROVIDER_READY',
-
-  /**
-   * The provider is in an error state.
-   */
-  Error = 'PROVIDER_ERROR',
-  
-  /**
-   * The flag configuration in the source-of-truth has changed.
-   */
-  ConfigurationChanged = 'PROVIDER_CONFIGURATION_CHANGED',
-  
-  /**
-   * The provider is transitioning to a state of unavailability.
-   */
-  Shutdown = 'PROVIDER_SHUTDOWN',
-};
-
-export interface EventData {
-  flagKeysChanged?: string[],
-  changeMetadata?: { [key: string]: boolean | string } // similar to flag metadata
-}
-
-export enum ApiEvents {
-  ProviderChanged = 'providerChanged',
-}
-
-export interface Eventing {
-  addHandler(notificationType: string, handler: Handler): void
-}
-
-export type EventContext = {
-  notificationType: string;
-  [key: string]: unknown;
-}
-
-export type Handler = (eventContext?: EventContext) => void
-
-export type EventCallbackMessage = (eventContext: EventContext) => void
-
 export type JsonObject = { [key: string]: JsonValue };
 
 export type JsonArray = JsonValue[];
